+++ packages/vue/src/useError.ts
@@ -1,9 +1,10 @@
-import { StateErrorTypes } from '@nhost/nhost-js'
+import { ErrorPayload, StateErrorTypes } from '@nhost/nhost-js'
 import { useSelector } from '@xstate/vue'
+import { Ref } from 'vue'
 import { useAuthInterpreter } from './useAuthInterpreter'
 
 /** @internal */
-export const useError = (type: StateErrorTypes) => {
+export const useError = (type: StateErrorTypes): Ref<ErrorPayload | null> => {
   const service = useAuthInterpreter()
   return useSelector(
     service.value,